[Maintenance] Move ProductSynchronizationRepositoryInterface
