+++ src/Repository/ProductSynchronizationRepositoryInterface.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace CommerceWeavers\SyliusAlsoBoughtPlugin\Doctrine\ORM;
+namespace CommerceWeavers\SyliusAlsoBoughtPlugin\Repository;
 
 use CommerceWeavers\SyliusAlsoBoughtPlugin\Entity\ProductSynchronizationInterface;
 use Sylius\Component\Resource\Repository\RepositoryInterface;
