+++ tests/Unit/Provider/LastSynchronizationDateProviderTest.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Tests\CommerceWeavers\SyliusAlsoBoughtPlugin\Unit\Provider;
 
-use CommerceWeavers\SyliusAlsoBoughtPlugin\Doctrine\ORM\ProductSynchronizationRepositoryInterface;
 use CommerceWeavers\SyliusAlsoBoughtPlugin\Entity\ProductSynchronization;
 use CommerceWeavers\SyliusAlsoBoughtPlugin\Provider\LastSynchronizationDateProvider;
+use CommerceWeavers\SyliusAlsoBoughtPlugin\Repository\ProductSynchronizationRepositoryInterface;
 use PHPUnit\Framework\TestCase;
 use Prophecy\PhpUnit\ProphecyTrait;
 use Symfony\Component\Uid\Uuid;
